Basic pong paddles working
The UI with some pong paddles and the ball are working. Still isn't as
smooth as I'd like it to be, although I have some ideas about fixing it.

+++ ICS4UC_RST/src/games/pong/pieces/PongPiece.java
@@ -4,10 +4,10 @@ package games.pong.pieces;
  * @author Kyle Anderson
  */
 public interface PongPiece {
-    int getX(Side position);
-    int getY(Side position);
-    void setX(int value, Side position);
-    void setY(int value, Side position);
-    int getHeight();
-    int getWidth();
+    double getX(Side position);
+    double getY(Side position);
+    void setX(double value, Side position);
+    void setY(double value, Side position);
+    double getHeight();
+    double getWidth();
 }
